Preval the theme file to drop polished from bundle
