+++ babel.config.js
@@ -6,6 +6,7 @@ function replacementPlugin(env) {
 
 const sharedPlugins = [
   'macros',
+  'preval',
   'add-react-displayname',
   'babel-plugin-styled-components',
   '@babel/plugin-proposal-object-rest-spread'
