+++ package.json
@@ -71,6 +71,7 @@
     "@rollup/plugin-node-resolve": "7.1.3",
     "babel-core": "7.0.0-bridge.0",
     "babel-plugin-add-react-displayname": "0.0.5",
+    "babel-plugin-preval": "5.0.0",
     "babel-plugin-styled-components": "1.10.7",
     "babel-plugin-transform-replace-expressions": "0.2.0",
     "enzyme": "3.10.0",
